apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
psycopg[binary,pool]==3.2.3
python-dotenv==1.0.1
langgraph==0.2.38
langchain==0.3.4
openai==1.51.2
supabase==2.6.0
numpy==1.26.4
tenacity==9.0.0

